backend/requirements.txt
Flask
spotipy
python-dotenv
textblob
